Add dynamic fills to results page
You guys can decide what you want to display, but I stuck with distance for the time being (I think that would be the ultimate result, although we currently have an arbitrary process of obtaining that number).  Note that the search query on main.js displays results where distance < 101, so tinker with that at will.